+++ packages/app/src/components-styled/tile.tsx
@@ -1,51 +1,14 @@
+import { Box } from './base';
 import { css } from '@styled-system/css';
 import styled from 'styled-components';
-import {
-  borders,
-  color,
-  flexbox,
-  grid,
-  layout,
-  position,
-  shadow,
-  space,
-  typography,
-  variant,
-} from 'styled-system';
-import { spacing } from '~/style/functions/spacing';
-import { BoxProps } from './base';
 
-interface TileProps extends BoxProps {
-  variant?: string;
-}
-
-export const Tile = styled.article<TileProps>(
+export const Tile = styled(Box).attrs({ as: 'article' })(
   css({
-    boxSizing: 'border-box',
-    minWidth: 0,
     display: 'flex',
     flexDirection: 'column',
+    bg: 'white',
     p: 4,
     borderRadius: 1,
     boxShadow: 'tile',
-    bg: 'white',
-  }),
-  variant({
-    variants: {
-      withoutBorder: css({
-        boxShadow: 'none',
-        p: 0,
-      }),
-    },
-  }),
-  space,
-  spacing,
-  layout,
-  flexbox,
-  color,
-  position,
-  typography,
-  borders,
-  shadow,
-  grid
+  })
 );
