+++ packages/app/src/domain/topical/mini-trend-tile-layout.tsx
@@ -1,6 +1,5 @@
 import { Children, ReactNode } from 'react';
 import { Box } from '~/components-styled/base';
-import { Tile } from '~/components-styled/tile';
 
 type MiniTrendTileLayoutProps = {
   children: ReactNode;
@@ -12,14 +11,12 @@ export function MiniTrendTileLayout({ children }: MiniTrendTileLayoutProps) {
   const columnWidth = `${Math.floor(100 / tiles.length)}%`;
 
   return (
-    <Tile variant="withoutBorder">
-      <Box display={{ _: 'block', md: 'flex' }}>
-        {tiles.map((tile: ReactNode, index: number) => (
-          <Box flex={`1 1 ${columnWidth}`} key={index}>
-            {tile}
-          </Box>
-        ))}
-      </Box>
-    </Tile>
+    <Box display={{ _: 'block', md: 'flex' }}>
+      {tiles.map((tile: ReactNode, index: number) => (
+        <Box flex={`1 1 ${columnWidth}`} key={index}>
+          {tile}
+        </Box>
+      ))}
+    </Box>
   );
 }
