+++ packages/app/src/pages/gemeente/[code]/actueel.tsx
@@ -16,7 +16,7 @@ import { escalationTooltip } from '~/components/choropleth/tooltips/region/escal
 import { FCWithLayout, getLayoutWithMetadata } from '~/domain/layout/layout';
 import { Search } from '~/domain/topical/components/search';
 import { DataSitemap } from '~/domain/topical/data-site-map';
-import { EscalationLevelExplanationsTile } from '~/domain/topical/escalation-level-explanations-tile';
+import { EscalationLevelExplanations } from '~/domain/topical/escalation-level-explanations';
 import { MiniTrendTile } from '~/domain/topical/mini-trend-tile';
 import { MiniTrendTileLayout } from '~/domain/topical/mini-trend-tile-layout';
 import { createGetStaticProps } from '~/static-props/create-get-static-props';
@@ -58,7 +58,7 @@ const MunicipalityActueel: FCWithLayout<typeof getStaticProps> = (props) => {
         <TileList>
           <MessageTile message={siteText.regionaal_index.belangrijk_bericht} />
 
-          <Search />
+          <Search initialValue={municipalityName} />
 
           <Heading level={1} fontWeight="normal">
             De actuele situatie in <strong>{municipalityName}</strong>
@@ -129,7 +129,7 @@ const MunicipalityActueel: FCWithLayout<typeof getStaticProps> = (props) => {
                 }),
               },
             ]}
-          ></QuickLinks>
+          />
 
           <ChoroplethTile
             title={text.risiconiveaus.selecteer_titel}
@@ -159,7 +159,7 @@ const MunicipalityActueel: FCWithLayout<typeof getStaticProps> = (props) => {
             />
           </ChoroplethTile>
 
-          <EscalationLevelExplanationsTile />
+          <EscalationLevelExplanations />
 
           <DataSitemap />
         </TileList>
